lab/github_slash_command/requirements.txt
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
Deprecated==1.2.15
idna==3.10
pycparser==2.22
PyGithub==2.4.0
PyJWT==2.9.0
PyNaCl==1.5.0
requests==2.32.3
structlog==24.4.0
typing_extensions==4.12.2
urllib3==2.2.3
wrapt==1.17.0
